+++ server/src-lib/Hasura/Backends/MSSQL/Instances/Execute.hs
@@ -10,6 +10,7 @@ import Control.Monad.Trans.Control (MonadBaseControl)
 import Control.Monad.Validate qualified as V
 import Data.Aeson.Extended qualified as J
 import Data.HashMap.Strict qualified as Map
+import Data.HashMap.Strict.InsOrd qualified as OMap
 import Data.HashSet qualified as Set
 import Data.List.NonEmpty qualified as NE
 import Data.Text.Extended qualified as T
@@ -26,6 +27,7 @@ import Hasura.Base.Error
 import Hasura.EncJSON
 import Hasura.GraphQL.Execute.Backend
 import Hasura.GraphQL.Execute.LiveQuery.Plan
+import Hasura.GraphQL.Namespace (RootFieldAlias (..), RootFieldMap)
 import Hasura.GraphQL.Parser
 import Hasura.Prelude
 import Hasura.RQL.IR
@@ -95,7 +97,7 @@ runShowplan query conn = do
 
 msDBQueryExplain ::
   MonadError QErr m =>
-  G.Name ->
+  RootFieldAlias ->
   UserInfo ->
   SourceName ->
   SourceConfig 'MSSQL ->
@@ -128,7 +130,7 @@ msDBLiveQueryExplain ::
   (MonadIO m, MonadBaseControl IO m, MonadError QErr m) =>
   LiveQueryPlan 'MSSQL (MultiplexedQuery 'MSSQL) ->
   m LiveQueryPlanExplanation
-msDBLiveQueryExplain (LiveQueryPlan plan sourceConfig variables) = do
+msDBLiveQueryExplain (LiveQueryPlan plan sourceConfig variables _) = do
   let (MultiplexedQuery' reselect) = _plqpQuery plan
       query = toQueryPretty $ fromSelect $ multiplexRootReselect [(dummyCohortId, variables)] reselect
       pool = _mscConnectionPool sourceConfig
@@ -433,15 +435,16 @@ msDBSubscriptionPlan ::
   UserInfo ->
   SourceName ->
   SourceConfig 'MSSQL ->
-  InsOrdHashMap G.Name (QueryDB 'MSSQL (Const Void) (UnpreparedValue 'MSSQL)) ->
+  Maybe G.Name ->
+  RootFieldMap (QueryDB 'MSSQL (Const Void) (UnpreparedValue 'MSSQL)) ->
   m (LiveQueryPlan 'MSSQL (MultiplexedQuery 'MSSQL))
-msDBSubscriptionPlan UserInfo {_uiSession, _uiRole} _sourceName sourceConfig rootFields = do
-  (reselect, prepareState) <- planSubscription rootFields _uiSession
+msDBSubscriptionPlan UserInfo {_uiSession, _uiRole} _sourceName sourceConfig namespace rootFields = do
+  (reselect, prepareState) <- planSubscription (OMap.mapKeys _rfaAlias rootFields) _uiSession
   cohortVariables <- prepareStateCohortVariables sourceConfig _uiSession prepareState
   let parameterizedPlan = ParameterizedLiveQueryPlan _uiRole $ MultiplexedQuery' reselect
 
   pure $
-    LiveQueryPlan parameterizedPlan sourceConfig cohortVariables
+    LiveQueryPlan parameterizedPlan sourceConfig cohortVariables namespace
 
 prepareStateCohortVariables :: (MonadError QErr m, MonadIO m, MonadBaseControl IO m) => SourceConfig 'MSSQL -> SessionVariables -> PrepareState -> m CohortVariables
 prepareStateCohortVariables sourceConfig session prepState = do